rename: DDD + Hexagonal 구조 변경
