+++ src/main/kotlin/uoslife/servermeeting/domain/meeting/domain/entity/MeetingTeam.kt
@@ -1,6 +1,7 @@
-package uoslife.servermeeting.entities
+package uoslife.servermeeting.domain.meeting.domain.entity
 
 import jakarta.persistence.*
+import uoslife.servermeeting.domain.match.domain.entity.Match
 
 @Entity
 @Table(name = "meeting_team")
