+++ src/main/kotlin/uoslife/servermeeting/domain/meeting/domain/entity/enums/TeamType.kt
@@ -1,0 +1,5 @@
+package uoslife.servermeeting.domain.meeting.domain.entity.enums
+
+enum class TeamType {
+    SINGLE, TRIPLE
+}
