+++ src/main/kotlin/uoslife/servermeeting/domain/user/domain/entity/enums/GenderType.kt
@@ -1,0 +1,5 @@
+package uoslife.servermeeting.domain.user.domain.entity.enums
+
+enum class GenderType {
+    MALE, FEMALE
+}
