--- src/main/kotlin/uoslife/servermeeting/entities/GenderType.kt
@@ -1,5 +1,0 @@
-package uoslife.servermeeting.entities
-
-enum class GenderType {
-    MALE, FEMALE
-}
