--- src/main/kotlin/uoslife/servermeeting/entities/TeamType.kt
@@ -1,5 +1,0 @@
-package uoslife.servermeeting.entities
-
-enum class TeamType {
-    SINGLE, TRIPLE
-}
